split code into Msgs_h.cs
