+++ Demo Plugin/NppManagedPluginDemo/NppManagedPluginDemo.VS2015.csproj
@@ -51,6 +51,9 @@
     <Compile Include="..\..\Visual Studio Project Template C#\Integration\Docking_h.cs">
       <Link>Integration\Docking_h.cs</Link>
     </Compile>
+    <Compile Include="..\..\Visual Studio Project Template C#\Integration\Msgs_h.cs">
+      <Link>Integration\Msgs_h.cs</Link>
+    </Compile>
     <Compile Include="..\..\Visual Studio Project Template C#\NppPluginNETBase.cs">
       <Link>NppPluginNETBase.cs</Link>
     </Compile>
