+++ Visual Studio Project Template C#/$projectname$.csproj
@@ -45,6 +45,7 @@
     <Compile Include="NppPluginNETBase.cs" />
     <Compile Include="NppPluginNETHelper.cs" />
     <Compile Include="Integration\Docking_h.cs" />
+    <Compile Include="Integration\Msgs_h.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
